+++ client/components/CostInputDownPayment.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { formatNum, parseUserStr } from '../utils';
-import { Box, Input, Slider } from './CostInputStyles.jsx';
+import { Box, Input, Slider } from './CostInput.styles.jsx';
 
 const Container = styled.span`
   width: 100%;
